fix(builder): fix migrations
