+++ libs/builder/schematics/migration.json
@@ -2,7 +2,7 @@
 	"$schema": "../../../node_modules/@angular-devkit/schematics/collection-schema.json",
 	"schematics": {
 		"migration-v17": {
-			"version": "17.0.0-beta.6",
+			"version": "17.0.0-beta.7",
 			"description": "Updates NgDoc to v17",
 			"factory": "./ng-update/index#update"
 		}
